web_api/requirements.txt
tornado==6.1
beautifulsoup4==4.10.0
stem==1.8
pysocks==1.7.1